Move code to OsxFileEventFunctions
