+++ src/main/java/net/rubygrapefruit/platform/internal/Platform.java
@@ -27,8 +27,8 @@ import net.rubygrapefruit.platform.file.FileSystems;
 import net.rubygrapefruit.platform.file.Files;
 import net.rubygrapefruit.platform.file.PosixFiles;
 import net.rubygrapefruit.platform.file.WindowsFiles;
-import net.rubygrapefruit.platform.internal.jni.DefaultOsxFileEventFunctions;
 import net.rubygrapefruit.platform.internal.jni.NativeLibraryFunctions;
+import net.rubygrapefruit.platform.internal.jni.OsxFileEventFunctions;
 import net.rubygrapefruit.platform.internal.jni.PosixTypeFunctions;
 import net.rubygrapefruit.platform.internal.jni.TerminfoFunctions;
 import net.rubygrapefruit.platform.memory.Memory;
@@ -346,8 +346,8 @@ public abstract class Platform {
             if (type.equals(Memory.class)) {
                 return type.cast(new DefaultMemory());
             }
-            if (type.equals(DefaultOsxFileEventFunctions.class)) {
-                return type.cast(new DefaultOsxFileEventFunctions());
+            if (type.equals(OsxFileEventFunctions.class)) {
+                return type.cast(new OsxFileEventFunctions());
             }
             return super.get(type, nativeLibraryLoader);
         }
